requirements.txt

streamlit
tensorflow
matplotlib
pillow
numpy
